my reduce kernels were incorrect

+++ headers/kernelDefs.h
@@ -51,9 +51,9 @@ __global__ void reduceFirstKern(kernelArray<double> weights,
                                 kernelArray<double> per_block_results,
                                 kernelArray<int> params, int device_offset);
 
-__global__ void reduceSecondKern(kernelArray<double> per_block_results, kernelArray<int> params, float *result);
+__global__ void reduceSecondKern(kernelArray<double> per_block_results, kernelArray<int> params, double *result);
 
-__global__ void normalizeKern(kernelArray<double> weights, kernelArray<int> params, float *avgFitness, int device_offset);
+__global__ void normalizeKern(kernelArray<double> weights, kernelArray<int> params, double *avgFitness, int device_offset);
 
 __global__ void evolutionKern(kernelArray<double> vect, kernelArray<int> params, int device_offset);
